replace lists with numpy arrays

+++ jtaglib.py
@@ -148,7 +148,7 @@ elif BACKEND=="ctypes":
         if (DataSize < 0 or DataSize >32):
             print("ERROR: Attempted to shift too large of data (>32 bits at a time).")
             sys.exit()
-        return(lptjtaglib.ShiftData(Data, DataSize, sendtms))
+        return(lptjtaglib.ShiftData(ctypes.c_int(Data), ctypes.c_int(DataSize), ctypes.c_int(sendtms)))
 
     #-------------------------------------------------------------------------------
     # Sends and receives JTAG data... wrapper to to Start IR/DR Shift, Shift in data
